Add tests for API functions, enhance chat navigation, and improve tool provider handling
- Implement unit tests for `listTools` and `saveTools` API functions to ensure correct behavior and error handling.
- Update `ChatScreen.svelte` to handle fresh chat requests via query parameters and improve state management.
- Modify `NavMain.svelte` to support button clicks for navigation and convert icons to emojis for better UX.
- Enhance `Sidebar.svelte` to include cache-busting query parameters for new chat navigation.
- Add comprehensive tests for web search rendering and tool provider metadata to ensure proper functionality.
- Set up Vitest configuration and testing environment for consistent testing practices across the codebase.

+++ crates/cleanclaw-auth/src/users.rs
@@ -334,7 +334,7 @@ impl Accounts {
             apikey_id: apikey_id.to_string(),
             external_id: external_id.to_string(),
             avatar_url: String::new(),
-            agent_quota: -1 as i32,
+            agent_quota: -1_i32,
             created_at: now,
             updated_at: now,
         };
